Feature/add configuration

+++ JPEG-Encoder/encoder.console/lib/Color.cs
@@ -2,15 +2,15 @@
 {
   public class Color
   {
-    public Color(double channel1, double channel2, double channel3)
+    public Color(float channel1, float channel2, float channel3)
     {
       Channel1 = channel1;
       Channel2 = channel2;
       Channel3 = channel3;
     }
 
-    public double Channel1 { get; set; }
-    public double Channel2 { get; set; }
-    public double Channel3 { get; set; }
+    public float Channel1 { get; set; }
+    public float Channel2 { get; set; }
+    public float Channel3 { get; set; }
   }
 }
